:recycle: Move datamodels, apis and messages into namespaces
