+++ src/data_models/index.ts
@@ -1,9 +1,20 @@
-export * from './correlation/index';
-export * from './event/index';
-export * from './kpi/index';
-export * from './logging/index';
-export * from './messages/index';
-export * from './process_model_execution/index';
-export * from './token_history/index';
-export * from './user_task/index';
-export * from './manual_task';
+import * as correlations from './correlation/index';
+import * as events from './event/index';
+import * as kpi from './kpi/index';
+import * as logging from './logging/index';
+import * as manualTasks from './manual_task';
+import * as processModels from './process_models/index';
+import * as tokenHistory from './token_history/index';
+import * as userTasks from './user_task/index';
+
+// tslint:disable-next-line:no-namespace
+export namespace DataModels {
+  export import Correlations = correlations;
+  export import Events = events;
+  export import Kpi = kpi;
+  export import Logging = logging;
+  export import ManualTasks = manualTasks;
+  export import ProcessModels = processModels;
+  export import TokenHistory = tokenHistory;
+  export import UserTasks = userTasks;
+}
