+++ src/data_models/process_models/process_model_list.ts
@@ -1,4 +1,4 @@
-import {ProcessModel} from '../index';
+import {ProcessModel} from './process_model';
 
 /**
  * Contains a list of ProcessModels.
