+++ src/index.ts
@@ -1,25 +1,9 @@
 
+export * from './apis/index';
+export * from './data_models/index';
+export * from './messages/index';
+
 export * from './imanagement_api';
 export * from './imanagement_api_accessor';
 export * from './rest_settings';
 export * from './socket_settings';
-
-export * from './data_models/correlation/index';
-export * from './data_models/event/index';
-export * from './data_models/kpi/index';
-export * from './data_models/logging/index';
-export * from './data_models/messages/index';
-export * from './data_models/token_history/index';
-export * from './data_models/user_task/index';
-export * from './data_models/manual_task/index';
-import * as execution from './data_models/process_model_execution/index';
-
-// tslint:disable-next-line:no-namespace
-export namespace ProcessModelExecution {
-  export import ProcessModel = execution.ProcessModel;
-  export import ProcessModelList = execution.ProcessModelList;
-  export import ProcessStartRequestPayload = execution.ProcessStartRequestPayload;
-  export import ProcessStartResponsePayload = execution.ProcessStartResponsePayload;
-  export import StartCallbackType = execution.StartCallbackType;
-  export import UpdateProcessDefinitionsRequestPayload = execution.UpdateProcessDefinitionsRequestPayload;
-}
